requirements.txt
intelhex~=2.2.1
hidapi~=0.9.0